small simplification around amount and weight values

+++ src/main/java/com/mobiquityinc/dto/BundleStats.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BundleStats {
-    private float currentWeight;
-    private float currentCost;
+    private double currentWeight;
+    private double currentCost;
 }
