+++ src/main/java/com/mobiquityinc/model/BasePackage.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public abstract class BasePackage implements IPackage {
-    private float maxWeightAllowed;
+    private double maxWeightAllowed;
     private List<Item> things;
 
     public abstract void addItemTo(Item item);
